# scripts/openapi.yaml
openapi: '3.1.0'
info:
  title: Rust Autorization Server
  description: Authorization Server written in rust using jwt/jws/jwe and featuring integration with Google OAuth
  contact:
    name: Daniel Boline
    email: ddboline@gmail.com
  license:
    name: MIT
    identifier: MIT
  version: '0.13.0'
paths:
...
  /api/session/{session_key}:
    get:
      summary: Get Session
      operationId: get_session
      responses:
        '200':
          description: Session Object
          content:
            application/json:
              schema: {}
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
        '401':
          description: Not Authorized
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
    post:
      summary: Set session value
      operationId: post_session
      requestBody:
        content:
          application/json:
            schema: {}
        required: true
      responses:
        '201':
          description: Set Session Object
          content:
            application/json:
              schema: {}
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
        '401':
          description: Not Authorized
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
    delete:
      summary: Delete session value
      operationId: delete_session
      responses:
        '204':
          description: Delete Session Object
          content:
            application/json:
              schema: {}
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
        '401':
          description: Not Authorized
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
...
  /auth/login.html:
    get:
      summary: Login Page
      operationId: login_html
      responses:
        '200':
          description: Login Page
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
        '401':
          description: Not Authorized
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                properties:
                  message:
                    $ref: '#/components/schemas/str'
...
components:
  schemas:
...
    str:
      type: string